Project- Squared Numbers
Created an app that loops between 1 and 100, the app prints out the numbers along side their squared value, and the app stops when a squared value of 100 or more is reached. To do: reconfigure the app to take user input and accept the value as well as printing it from 1 to the user inputted value.